refactor separated list; add tests
